requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
Pillow==10.0.0
pytesseract==0.3.10
opencv-python==4.8.0.76
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
structlog==23.1.0
gunicorn==21.2.0
